Fix: Only whitelist when sending email (#3145)
* fix(only-whitelist-when-sending-email): Move whitelist to sendEmail logic

* fix(only-whitelist-when-sending-email): Remove whitelist from nextEmail
